Move strategies in `gen` package
