+++ src/main/java/one/nio/serial/gen/MagicAccessor.java
@@ -20,7 +20,7 @@ import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
-import static one.nio.serial.gen.strategy.MagicAccessorStrategy.MAGIC_CLASS;
+import static one.nio.serial.gen.MagicAccessorStrategy.MAGIC_CLASS;
 
 public class MagicAccessor {
